requirements.txt
pyspellchecker==23.2.1